OHRM5X-1454: moved review evaluate code into admin evaluate
