+++ src/client/src/orangehrmPerformancePlugin/index.ts
@@ -31,7 +31,7 @@ import ReviewList from './pages/ReviewList.vue';
 import ReviewSearch from './pages/ReviewSearch.vue';
 import EmployeeTrackerLogs from './pages/EmployeeTrackerLogs.vue';
 import AdminEvaluation from './pages/AdminEvaluation.vue';
-import SelfReviewEvaluate from './pages/SelfReviewEvaluate.vue';
+import SelfEvaluation from './pages/SelfEvaluation.vue';
 
 export default {
   'my-tracker': MyTracker,
@@ -49,5 +49,5 @@ export default {
   'review-search': ReviewSearch,
   'employee-tracker-logs': EmployeeTrackerLogs,
   'admin-evaluation': AdminEvaluation,
-  'self-review-evaluate': SelfReviewEvaluate,
+  'self-evaluation': SelfEvaluation,
 };
