--- src/plugins/orangehrmPerformancePlugin/Controller/ReviewEvaluateController.php
@@ -1,98 +1,0 @@
-<?php
-/**
- * OrangeHRM is a comprehensive Human Resource Management (HRM) System that captures
- * all the essential functionalities required for any enterprise.
- * Copyright (C) 2006 OrangeHRM Inc., http://www.orangehrm.com
- *
- * OrangeHRM is free software; you can redistribute it and/or modify it under the terms of
- * the GNU General Public License as published by the Free Software Foundation; either
- * version 2 of the License, or (at your option) any later version.
- *
- * OrangeHRM is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along with this program;
- * if not, write to the Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor,
- * Boston, MA  02110-1301, USA
- */
-
-namespace OrangeHRM\Performance\Controller;
-
-use OrangeHRM\Core\Authorization\Controller\CapableViewController;
-use OrangeHRM\Core\Controller\AbstractVueController;
-use OrangeHRM\Core\Controller\Common\NoRecordsFoundController;
-use OrangeHRM\Core\Controller\Exception\RequestForwardableException;
-use OrangeHRM\Core\Traits\UserRoleManagerTrait;
-use OrangeHRM\Core\Vue\Component;
-use OrangeHRM\Core\Vue\Prop;
-use OrangeHRM\Entity\PerformanceReview;
-use OrangeHRM\Framework\Http\Request;
-use OrangeHRM\Performance\Traits\Service\PerformanceReviewServiceTrait;
-
-class ReviewEvaluateController extends AbstractVueController implements CapableViewController
-{
-    use PerformanceReviewServiceTrait;
-    use UserRoleManagerTrait;
-
-    /**
-     * @inheritDoc
-     */
-    public function preRender(Request $request): void
-    {
-        $id = $request->attributes->getInt('id');
-        $component = new Component('review-evaluate');
-
-        $review = $this->getPerformanceReviewService()->getPerformanceReviewDao()->getPerformanceReviewById($id);
-        if (!is_null($review)) {
-            $this->setReviewProps($component, $review);
-            if ($this->isUserPerformanceReviewEvaluator($id)) {
-                $component->addProp(new Prop('is-reviewer', Prop::TYPE_BOOLEAN, true));
-            }
-        }
-        $this->setComponent($component);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function isCapable(Request $request): bool
-    {
-        $id = $request->attributes->getInt('id');
-        if (is_null($this->getPerformanceReviewService()->getPerformanceReviewDao()->getPerformanceReviewById($id))) {
-            throw new RequestForwardableException(NoRecordsFoundController::class . '::handle');
-        }
-        return $this->getUserRoleManager()->isEntityAccessible(PerformanceReview::class, $id, null, ['ESS']);
-    }
-
-    /**
-     * @param Component $component
-     * @param PerformanceReview $performanceReview
-     */
-    protected function setReviewProps(Component $component, PerformanceReview $performanceReview): void
-    {
-        $component->addProp(new Prop('review-id', Prop::TYPE_NUMBER, $performanceReview->getId()));
-        $component->addProp(new Prop('emp-number', Prop::TYPE_NUMBER, $performanceReview->getEmployee()->getEmpNumber()));
-        $component->addProp(new Prop('employee-name', Prop::TYPE_STRING, $performanceReview->getEmployee()->getDecorator()->getFirstAndLastNames()));
-        $component->addProp(new Prop('job-title', Prop::TYPE_STRING, $performanceReview->getEmployee()->getJobTitle()->getJobTitleName()));
-        $component->addProp(new Prop('status', Prop::TYPE_NUMBER, $performanceReview->getStatusId()));
-        $component->addProp(new Prop('review-period-start', Prop::TYPE_STRING, $performanceReview->getDecorator()->getReviewPeriodStart()));
-        $component->addProp(new Prop('review-period-end', Prop::TYPE_STRING, $performanceReview->getDecorator()->getReviewPeriodEnd()));
-        $component->addProp(new Prop('due-date', Prop::TYPE_STRING, $performanceReview->getDecorator()->getDueDate()));
-    }
-
-    /**
-     * @param int $performanceReviewId
-     * @return bool
-     */
-    private function isUserPerformanceReviewEvaluator(int $performanceReviewId): bool
-    {
-        return $this->getUserRoleManager()->isEntityAccessible(
-            PerformanceReview::class,
-            $performanceReviewId,
-            null,
-            ['Admin', 'ESS'],
-            ['Supervisor']
-        );
-    }
-}
